iter9 - moved all code to internal
